Split up mark calculator into two modules
